Add files via upload
The excel files added in this upload are necessary in order to run the model. A Course Case file has been added as well for readers to understand better the business context of this analysis. The Analysis and Code file contains a detailed explanation of the model and the steps necessary to extract valuable insights. I would advise readers to go through the Analysis and Code file on Jupyter Notebook before reading the Analysis Write-Up. The Analysis Write-Up file is a summary of the insights as well as the proposed recommendations based on results generated from the model. Finally, the final model is a .py file that contains only the most successful classification model used to present the insights and recommendation.